vyos-gce-image/chroot-patches/opt/gce_helper/requirements.txt
wheel
google-cloud-pubsub
google-api-python-client
google-cloud-storage
google-cloud-compute
google-cloud-logging